refactor(socratic): split socratic_agent.py into focused modules (Related to #1829)
785-line monolith → 5 focused modules under app/services/socratic/:
- models.py: SessionState dataclass (24 lines)
- session_store.py: SessionStore with DB-backed cache + rate limit (170 lines)
- prompt_builder.py: SocraticPromptBuilder with phase-aware prompt construction (143 lines)
- state_machine.py: PHASE_ORDER, determine_phase, rebuild_state_from_turns (61 lines)
- __init__.py: SocraticAgent facade + AgentResponse + _fallback_question (444 lines)

socratic_agent.py becomes 2-line compatibility shim. All callers unchanged.
Smoke tests pass: shim imports + FastAPI app builds.
Circuit breaker, rate limit, fail-closed, and reasoning field all preserved.
