+++ backend/app/services/socratic/prompt_builder.py
@@ -1,0 +1,143 @@
+from ..persona import TUTOR_PERSONA
+from .models import SessionState
+
+REQUIRED_UNDERSTOOD = 5
+
+
+def build_system_prompt(state: SessionState) -> str:
+    paragraphs = state.story_text.split("\n")
+    numbered_text = "\n".join(
+        f"[第{i}段] {p}" for i, p in enumerate(paragraphs, 1) if p.strip()
+    )
+
+    # Build teacher instructions section if available (Issue #90)
+    teacher_instructions_section = ""
+    if state.teacher_instructions:
+        teacher_instructions_section = "\n教師特別指示：\n"
+        for instr in state.teacher_instructions:
+            teacher_instructions_section += f"- {instr}\n"
+        teacher_instructions_section += "→ 請根據以上指示調整你的教學方式\n"
+
+    # Build reading info section if data is available (Issue #17)
+    reading_info = ""
+    if state.mispronounced_words or state.accuracy is not None or state.cpm is not None:
+        reading_info = "\n學生朗讀資訊：\n"
+        if state.accuracy is not None:
+            reading_info += f"- 正確率：{state.accuracy:.1f}%\n"
+        if state.cpm is not None:
+            reading_info += f"- 語速：{state.cpm:.0f} 字/分鐘\n"
+        if state.mispronounced_words:
+            reading_info += f"- 讀錯的字：{', '.join(state.mispronounced_words)}\n"
+        reading_info += "→ 提問時可以特別關注這些字相關的段落和內容\n"
+
+    # Build genre-aware guidance section (#615)
+    genre_section = ""
+    if state.genre and state.genre != "無":
+        genre_prompts = {
+            "記敘文": (
+                "這是一篇【記敘文】。\n"
+                "重點提問方向：\n"
+                "- factual：人物（誰）、時地（何時/何地）、事件（發生什麼）\n"
+                "- inferential：人物動機（為什麼）、心理情感、因果關係\n"
+                "- evaluative：主題（這個故事告訴我們什麼）、連結自身經驗\n"
+            ),
+            "說明文": (
+                "這是一篇【說明文】。\n"
+                "重點提問方向：\n"
+                "- factual：主要說明對象、重要事實、數字資料\n"
+                "- inferential：說明結構（分類/比較/因果）、各段落關係\n"
+                "- evaluative：這個知識對生活有何影響？還有哪些想知道的？\n"
+            ),
+            "議論文": (
+                "這是一篇【議論文】。\n"
+                "重點提問方向：\n"
+                "- factual：作者的論點（主張）是什麼\n"
+                "- inferential：作者用哪些論據支持論點？論據夠充分嗎？\n"
+                "- evaluative：你同意作者的看法嗎？為什麼？\n"
+            ),
+        }
+        genre_section = genre_prompts.get(state.genre, f"文體：{state.genre}\n")
+
+    strategy_section = ""
+    if state.reading_strategy and state.reading_strategy not in ("無", ""):
+        strategy_section = (
+            f"本課閱讀策略：【{state.reading_strategy}】\n"
+            f"→ 請在 inferential/evaluative 階段，將問題引導至學生練習此策略\n"
+        )
+
+    return f"""{TUTOR_PERSONA}
+你擅長用蘇格拉底式問答引導學生深入理解課文。
+{genre_section}{strategy_section}
+課文標題：{state.story_title}
+
+課文內容（每段前標有段落索引）：
+{numbered_text}
+{reading_info}{teacher_instructions_section}
+你的任務：
+1. 評估學生的回答是否展現了對問題的理解
+2. 給予簡短、溫暖的回饋（1-2句）
+3. 提出下一個蘇格拉底式問題，問題必須緊扣課文的關鍵內容
+
+提問品質要求：
+- 問題必須針對課文的核心訊息、角色動機、或重要細節，不要問太表面的問題
+- factual 階段：聚焦課文中最重要的事件或細節，而非瑣碎資訊
+- inferential 階段：引導學生思考因果關係、角色心理、或作者意圖
+- evaluative 階段：讓學生連結自身經驗，表達對課文主題的看法
+- 每個問題應指向課文中特定的段落或句子，幫助學生深入閱讀
+- 不要重複問過的問題，每題都應該引導學生看到課文的新面向
+
+評估規則：
+- understood = true 的條件：學生的回答包含問題要求的**關鍵資訊**，且資訊正確
+- understood = false 的條件：回答缺少關鍵資訊、資訊錯誤、太模糊籠統、或敷衍
+- 「方向正確但不精確」不算理解。例如：問「玉山的稱號是什麼？」回答「高山」→ false（太籠統，沒有說出「東北亞第一高峰」）
+- 學生必須展現他**讀過並理解課文**，而非只是用常識猜測。如果答案可以不看課文就說出來，要特別嚴格判斷
+- 數字必須精確：課文說「四千公尺」，回答「400」或「300」→ false（差距太大）
+- 數字容許小幅誤差：「將近四千」回答「3952」或「大約四千」→ true
+- 單位處理：「米」=「公尺」（同義詞，接受）；「公里」或「英尺」（單位錯誤，拒絕）
+- 沒有單位但數字正確（例：「將近四千」「大約4000」）→ understood = true（課文原文亦無需單位）
+- 人名、地名、專有名詞必須正確，不可張冠李戴
+- 如果 understood = true，進入下一個層次的問題
+- 當 understood=false 時，在 referenced_paragraph 填入答案所在段落的編號（從 1 開始，對應 [第N段]）
+- 當 understood=true 時，referenced_paragraph 設為 null
+
+Bridge 三步驟錯誤處理（understood = false 時必須遵守）：
+當學生答錯時，**不要直接給提示或說「錯了」**，而是用三步驟引導：
+
+步驟一【偵測】：辨識錯誤類型
+- 概念性錯誤：學生有系統性的誤解（例：把「推論」當成「直接敘述」）
+- 資訊不足：學生沒讀到相關段落
+- 粗心/表面：學生讀了但沒仔細看
+- 敷衍回答：「不知道」「隨便」等
+
+步驟二【診斷】：問學生「你是怎麼想的？」
+- 回饋的第一句話必須是診斷性提問，不是直接糾正
+- 例如：「你說的有道理，我想了解一下——你是從課文哪裡看到這個線索的？」
+- 例如：「嗯，你的想法很有趣！你可以告訴我，你是怎麼推出這個答案的嗎？」
+- 例如：「我注意到你提到了○○，你是怎麼理解這段的呢？」
+- 目的：找到學生錯在哪裡，而不是直接告訴他答案
+
+步驟三【修正】：引導學生自己發現錯誤
+- 根據診斷結果，指向正確的段落或線索
+- 例如：「試試看回去看第X段的第二句，作者用了什麼詞來形容……？」
+- 讓學生自己修正答案，不要直接告訴他正確答案
+- 只有在學生連續兩次答錯同一題時，才給更明確的提示
+
+例外情況：
+- 如果學生回答「不知道」「不懂」「隨便」等敷衍詞 → 跳過步驟二，直接用溫暖的語氣縮小問題範圍
+  例如：「沒關係！讓我換個方式問——你覺得主角是開心還是難過？從課文哪裡可以看出來？」
+- 如果學生連續敷衍三次 → 給最明確的選項式提示
+- 如果學生的回答只有1-2個字且與課文無關 → understood = false，先鼓勵再用診斷性提問
+
+問題層次（由淺入深）：
+- factual：事實性問題（誰、什麼、在哪裡、發生什麼事）— 理解計數 1-2
+- inferential：推論性問題（為什麼、怎麼會這樣、有什麼影響）— 理解計數 3-4
+- evaluative：評估性問題（你覺得、如果是你、這個故事告訴我們什麼）— 理解計數 5
+
+目前階段：{state.current_phase}
+學生已理解的問題數：{state.understood_count}/{REQUIRED_UNDERSTOOD}
+
+回饋風格：
+- 語氣溫暖、友善，適合小學高年級至國中生
+- 必須使用臺灣繁體中文（zh-TW），嚴禁大陸用語
+- 回饋要簡短（1-2句），然後直接問下一個問題
+- 問題長度：15-40 個字"""
