+++ backend/app/services/socratic_agent.py
@@ -1,785 +1,2 @@
-"""
-Socratic dialogue agent with answer evaluation.
-Manages session state, evaluates student comprehension, and generates
-follow-up questions using structured Gemini output.
-"""
-
-import time
-import logging
-from dataclasses import dataclass, field
-
-from google.genai import types as genai_types
-from sqlalchemy.exc import SQLAlchemyError
-from sqlalchemy.orm import Session
-
-from .ai_service import generate_structured_response
-from .input_sanitizer import sanitize_ai_input
-from .persona import TUTOR_PERSONA
-
-logger = logging.getLogger(__name__)
-
-
-@dataclass
-class SessionState:
-    session_id: str
-    story_title: str
-    story_text: str
-    conversation: list[dict] = field(default_factory=list)  # {role, text}
-    understood_count: int = 0
-    total_attempts: int = 0
-    current_phase: str = "factual"  # factual -> inferential -> evaluative
-    created_at: float = field(default_factory=time.time)
-    consecutive_errors: int = 0
-    # Optional reading results from LiveTutor (Issue #17)
-    mispronounced_words: list[str] | None = None
-    accuracy: float | None = None
-    cpm: float | None = None
-    # Teacher special instructions for individualized learning (Issue #90)
-    teacher_instructions: list[str] | None = None
-    # Genre-aware Socratic (#615)
-    genre: str | None = None
-    reading_strategy: str | None = None
-
-
-class SessionStore:
-    """DB-backed session store with in-memory hot cache and 30-minute TTL cleanup.
-
-    Persistence strategy (Issue #961):
-    - save(state, db, db_session_id): writes serialized SessionState to
-      LearningSession.dialogue_state JSONB column AND the in-memory cache.
-    - get(session_id, db, db_session_id): reads from in-memory cache first
-      (fast path); falls back to DB on cache miss (survives Cloud Run restarts).
-    - save(state) without db: writes to in-memory cache only (backwards compatible).
-    - get(session_id) without db: returns None on cache miss (backwards compatible).
-    """
-
-    TTL_SECONDS = 30 * 60
-    RATE_LIMIT = 30  # max requests per minute per session
-    RATE_WINDOW = 60  # seconds
-
-    def __init__(self):
-        self._sessions: dict[str, SessionState] = {}
-        self._rate_counts: dict[str, list[float]] = {}  # session_id -> [timestamps]
-
-    def get(
-        self,
-        session_id: str,
-        db: "Session | None" = None,
-        db_session_id: int | None = None,
-    ) -> SessionState | None:
-        """Return SessionState from in-memory cache, or restore from DB on miss."""
-        self._cleanup()
-        if session_id in self._sessions:
-            return self._sessions[session_id]
-
-        # Cache miss — attempt DB restore if context provided
-        if db is not None and db_session_id is not None:
-            return self._restore_from_db(session_id, db, db_session_id)
-
-        return None
-
-    def save(
-        self,
-        state: SessionState,
-        db: "Session | None" = None,
-        db_session_id: int | None = None,
-    ) -> None:
-        """Persist SessionState to in-memory cache, and optionally to DB."""
-        self._sessions[state.session_id] = state
-
-        if db is not None and db_session_id is not None:
-            self._persist_to_db(state, db, db_session_id)
-
-    # ------------------------------------------------------------------
-    # Internal helpers
-    # ------------------------------------------------------------------
-
-    def _serialize(self, state: SessionState) -> dict:
-        """Convert SessionState dataclass to a JSON-serializable dict."""
-        import dataclasses
-        return dataclasses.asdict(state)
-
-    def _deserialize(self, data: dict) -> SessionState:
-        """Reconstruct SessionState from a dict (e.g., loaded from JSONB)."""
-        return SessionState(
-            session_id=data["session_id"],
-            story_title=data.get("story_title", ""),
-            story_text=data.get("story_text", ""),
-            conversation=data.get("conversation", []),
-            understood_count=data.get("understood_count", 0),
-            total_attempts=data.get("total_attempts", 0),
-            current_phase=data.get("current_phase", "factual"),
-            created_at=data.get("created_at", time.time()),
-            consecutive_errors=data.get("consecutive_errors", 0),
-            mispronounced_words=data.get("mispronounced_words"),
-            accuracy=data.get("accuracy"),
-            cpm=data.get("cpm"),
-            teacher_instructions=data.get("teacher_instructions"),
-            genre=data.get("genre"),
-            reading_strategy=data.get("reading_strategy"),
-        )
-
-    def _persist_to_db(
-        self,
-        state: SessionState,
-        db: "Session",
-        db_session_id: int,
-    ) -> None:
-        """Write serialized state to LearningSession.dialogue_state column."""
-        try:
-            from ..models.session import LearningSession  # noqa: PLC0415
-
-            ls = db.query(LearningSession).filter_by(id=db_session_id).first()
-            if ls is not None:
-                ls.dialogue_state = self._serialize(state)
-                db.commit()
-        except SQLAlchemyError:
-            logger.warning(
-                "Failed to persist dialogue_state to DB for db_session_id=%s",
-                db_session_id,
-                exc_info=True,
-            )
-            try:
-                db.rollback()
-            except SQLAlchemyError:
-                pass
-
-    def _restore_from_db(
-        self,
-        session_id: str,
-        db: "Session",
-        db_session_id: int,
-    ) -> SessionState | None:
-        """Load SessionState from LearningSession.dialogue_state column."""
-        try:
-            from ..models.session import LearningSession  # noqa: PLC0415
-
-            ls = db.query(LearningSession).filter_by(id=db_session_id).first()
-            if ls is None or ls.dialogue_state is None:
-                return None
-
-            data = ls.dialogue_state
-            # Validate that the stored snapshot belongs to this session
-            if data.get("session_id") != session_id:
-                logger.warning(
-                    "dialogue_state session_id mismatch: stored=%s requested=%s",
-                    data.get("session_id"),
-                    session_id,
-                )
-                return None
-
-            state = self._deserialize(data)
-            # Warm the in-memory cache so subsequent calls don't hit DB
-            self._sessions[session_id] = state
-            return state
-        except (SQLAlchemyError, AttributeError, KeyError):
-            logger.warning(
-                "Failed to restore dialogue_state from DB for db_session_id=%s",
-                db_session_id,
-                exc_info=True,
-            )
-            return None
-
-    def _cleanup(self) -> None:
-        now = time.time()
-        expired = [
-            sid for sid, s in self._sessions.items()
-            if now - s.created_at > self.TTL_SECONDS
-        ]
-        for sid in expired:
-            del self._sessions[sid]
-
-    def check_rate_limit(self, session_id: str) -> bool:
-        """Return True if rate limit exceeded."""
-        now = time.time()
-        timestamps = self._rate_counts.get(session_id, [])
-        # Remove old timestamps
-        timestamps = [t for t in timestamps if now - t < self.RATE_WINDOW]
-        if len(timestamps) >= self.RATE_LIMIT:
-            return True
-        timestamps.append(now)
-        self._rate_counts[session_id] = timestamps
-        return False
-
-
-# Module-level singleton
-_store = SessionStore()
-
-
-# Gemini response schema for structured output
-EVALUATION_SCHEMA = {
-    "type": "object",
-    "properties": {
-        "understood": {
-            "type": "boolean",
-            "description": "Whether the student's answer demonstrates understanding of the question",
-        },
-        "feedback": {
-            "type": "string",
-            "description": "Brief, warm feedback on the student's answer (1-2 sentences in Traditional Chinese)",
-        },
-        "question": {
-            "type": "string",
-            "description": "The next Socratic question in Traditional Chinese",
-        },
-        "phase": {
-            "type": "string",
-            "enum": ["factual", "inferential", "evaluative"],
-            "description": "The phase of the next question",
-        },
-        "referenced_paragraph": {
-            "type": "INTEGER",
-            "description": "答錯時，答案所在的段落編號（從 1 開始）。答對時為 null。",
-            "nullable": True,
-        },
-    },
-    "required": ["understood", "feedback", "question", "phase"],
-}
-
-PHASE_ORDER = ["factual", "inferential", "evaluative"]
-
-
-@dataclass
-class AgentResponse:
-    question: str
-    feedback: str | None
-    understood: bool | None
-    understood_count: int
-    required_count: int
-    phase: str
-    is_complete: bool
-    referenced_paragraph: int | None = None
-
-
-class SocraticAgent:
-    REQUIRED_UNDERSTOOD = 5
-    MAX_ANSWER_LENGTH = 500
-    MAX_HISTORY_TURNS = 10
-    MAX_CONSECUTIVE_ERRORS = 3
-
-    def _build_system_prompt(self, state: SessionState) -> str:
-        paragraphs = state.story_text.split("\n")
-        numbered_text = "\n".join(
-            f"[第{i}段] {p}" for i, p in enumerate(paragraphs, 1) if p.strip()
-        )
-
-        # Build teacher instructions section if available (Issue #90)
-        teacher_instructions_section = ""
-        if state.teacher_instructions:
-            teacher_instructions_section = "\n教師特別指示：\n"
-            for instr in state.teacher_instructions:
-                teacher_instructions_section += f"- {instr}\n"
-            teacher_instructions_section += "→ 請根據以上指示調整你的教學方式\n"
-
-        # Build reading info section if data is available (Issue #17)
-        reading_info = ""
-        if state.mispronounced_words or state.accuracy is not None or state.cpm is not None:
-            reading_info = "\n學生朗讀資訊：\n"
-            if state.accuracy is not None:
-                reading_info += f"- 正確率：{state.accuracy:.1f}%\n"
-            if state.cpm is not None:
-                reading_info += f"- 語速：{state.cpm:.0f} 字/分鐘\n"
-            if state.mispronounced_words:
-                reading_info += f"- 讀錯的字：{', '.join(state.mispronounced_words)}\n"
-            reading_info += "→ 提問時可以特別關注這些字相關的段落和內容\n"
-
-        # Build genre-aware guidance section (#615)
-        genre_section = ""
-        if state.genre and state.genre != "無":
-            genre_prompts = {
-                "記敘文": (
-                    "這是一篇【記敘文】。\n"
-                    "重點提問方向：\n"
-                    "- factual：人物（誰）、時地（何時/何地）、事件（發生什麼）\n"
-                    "- inferential：人物動機（為什麼）、心理情感、因果關係\n"
-                    "- evaluative：主題（這個故事告訴我們什麼）、連結自身經驗\n"
-                ),
-                "說明文": (
-                    "這是一篇【說明文】。\n"
-                    "重點提問方向：\n"
-                    "- factual：主要說明對象、重要事實、數字資料\n"
-                    "- inferential：說明結構（分類/比較/因果）、各段落關係\n"
-                    "- evaluative：這個知識對生活有何影響？還有哪些想知道的？\n"
-                ),
-                "議論文": (
-                    "這是一篇【議論文】。\n"
-                    "重點提問方向：\n"
-                    "- factual：作者的論點（主張）是什麼\n"
-                    "- inferential：作者用哪些論據支持論點？論據夠充分嗎？\n"
-                    "- evaluative：你同意作者的看法嗎？為什麼？\n"
-                ),
-            }
-            genre_section = genre_prompts.get(state.genre, f"文體：{state.genre}\n")
-
-        strategy_section = ""
-        if state.reading_strategy and state.reading_strategy not in ("無", ""):
-            strategy_section = (
-                f"本課閱讀策略：【{state.reading_strategy}】\n"
-                f"→ 請在 inferential/evaluative 階段，將問題引導至學生練習此策略\n"
-            )
-
-        return f"""{TUTOR_PERSONA}
-你擅長用蘇格拉底式問答引導學生深入理解課文。
-{genre_section}{strategy_section}
-課文標題：{state.story_title}
-
-課文內容（每段前標有段落索引）：
-{numbered_text}
-{reading_info}{teacher_instructions_section}
-你的任務：
-1. 評估學生的回答是否展現了對問題的理解
-2. 給予簡短、溫暖的回饋（1-2句）
-3. 提出下一個蘇格拉底式問題，問題必須緊扣課文的關鍵內容
-
-提問品質要求：
-- 問題必須針對課文的核心訊息、角色動機、或重要細節，不要問太表面的問題
-- factual 階段：聚焦課文中最重要的事件或細節，而非瑣碎資訊
-- inferential 階段：引導學生思考因果關係、角色心理、或作者意圖
-- evaluative 階段：讓學生連結自身經驗，表達對課文主題的看法
-- 每個問題應指向課文中特定的段落或句子，幫助學生深入閱讀
-- 不要重複問過的問題，每題都應該引導學生看到課文的新面向
-
-評估規則：
-- understood = true 的條件：學生的回答包含問題要求的**關鍵資訊**，且資訊正確
-- understood = false 的條件：回答缺少關鍵資訊、資訊錯誤、太模糊籠統、或敷衍
-- 「方向正確但不精確」不算理解。例如：問「玉山的稱號是什麼？」回答「高山」→ false（太籠統，沒有說出「東北亞第一高峰」）
-- 學生必須展現他**讀過並理解課文**，而非只是用常識猜測。如果答案可以不看課文就說出來，要特別嚴格判斷
-- 數字必須精確：課文說「四千公尺」，回答「400」或「300」→ false（差距太大）
-- 數字容許小幅誤差：「將近四千」回答「3952」或「大約四千」→ true
-- 單位處理：「米」=「公尺」（同義詞，接受）；「公里」或「英尺」（單位錯誤，拒絕）
-- 沒有單位但數字正確（例：「將近四千」「大約4000」）→ understood = true（課文原文亦無需單位）
-- 人名、地名、專有名詞必須正確，不可張冠李戴
-- 如果 understood = true，進入下一個層次的問題
-- 當 understood=false 時，在 referenced_paragraph 填入答案所在段落的編號（從 1 開始，對應 [第N段]）
-- 當 understood=true 時，referenced_paragraph 設為 null
-
-Bridge 三步驟錯誤處理（understood = false 時必須遵守）：
-當學生答錯時，**不要直接給提示或說「錯了」**，而是用三步驟引導：
-
-步驟一【偵測】：辨識錯誤類型
-- 概念性錯誤：學生有系統性的誤解（例：把「推論」當成「直接敘述」）
-- 資訊不足：學生沒讀到相關段落
-- 粗心/表面：學生讀了但沒仔細看
-- 敷衍回答：「不知道」「隨便」等
-
-步驟二【診斷】：問學生「你是怎麼想的？」
-- 回饋的第一句話必須是診斷性提問，不是直接糾正
-- 例如：「你說的有道理，我想了解一下——你是從課文哪裡看到這個線索的？」
-- 例如：「嗯，你的想法很有趣！你可以告訴我，你是怎麼推出這個答案的嗎？」
-- 例如：「我注意到你提到了○○，你是怎麼理解這段的呢？」
-- 目的：找到學生錯在哪裡，而不是直接告訴他答案
-
-步驟三【修正】：引導學生自己發現錯誤
-- 根據診斷結果，指向正確的段落或線索
-- 例如：「試試看回去看第X段的第二句，作者用了什麼詞來形容……？」
-- 讓學生自己修正答案，不要直接告訴他正確答案
-- 只有在學生連續兩次答錯同一題時，才給更明確的提示
-
-例外情況：
-- 如果學生回答「不知道」「不懂」「隨便」等敷衍詞 → 跳過步驟二，直接用溫暖的語氣縮小問題範圍
-  例如：「沒關係！讓我換個方式問——你覺得主角是開心還是難過？從課文哪裡可以看出來？」
-- 如果學生連續敷衍三次 → 給最明確的選項式提示
-- 如果學生的回答只有1-2個字且與課文無關 → understood = false，先鼓勵再用診斷性提問
-
-問題層次（由淺入深）：
-- factual：事實性問題（誰、什麼、在哪裡、發生什麼事）— 理解計數 1-2
-- inferential：推論性問題（為什麼、怎麼會這樣、有什麼影響）— 理解計數 3-4
-- evaluative：評估性問題（你覺得、如果是你、這個故事告訴我們什麼）— 理解計數 5
-
-目前階段：{state.current_phase}
-學生已理解的問題數：{state.understood_count}/{self.REQUIRED_UNDERSTOOD}
-
-回饋風格：
-- 語氣溫暖、友善，適合小學高年級至國中生
-- 必須使用臺灣繁體中文（zh-TW），嚴禁大陸用語
-- 回饋要簡短（1-2句），然後直接問下一個問題
-- 問題長度：15-40 個字"""
-
-    async def start_session(
-        self,
-        session_id: str,
-        story_title: str,
-        story_text: str,
-        mispronounced_words: list[str] | None = None,
-        accuracy: float | None = None,
-        cpm: float | None = None,
-        teacher_instructions: list[str] | None = None,
-        genre: str | None = None,
-        reading_strategy: str | None = None,
-        db: Session | None = None,
-        db_session_id: int | None = None,
-    ) -> tuple["AgentResponse", bool]:
-        """Start or resume a session.
-
-        Returns (AgentResponse, is_resumed).
-        is_resumed=True means the session was restored from memory or DB (no Gemini call).
-        is_resumed=False means a fresh Gemini call was made.
-        """
-        # 1. Check in-memory cache first; fall back to dialogue_state DB snapshot (Issue #961)
-        existing_state = _store.get(session_id, db=db, db_session_id=db_session_id)
-        if existing_state is not None:
-            last_ai_turn = next(
-                (t for t in reversed(existing_state.conversation) if t["role"] == "ai"),
-                None,
-            )
-            question = last_ai_turn["text"] if last_ai_turn else "請繼續回答上一個問題。"
-            return AgentResponse(
-                question=question,
-                feedback=None,
-                understood=None,
-                understood_count=existing_state.understood_count,
-                required_count=self.REQUIRED_UNDERSTOOD,
-                phase=existing_state.current_phase,
-                is_complete=existing_state.understood_count >= self.REQUIRED_UNDERSTOOD,
-            ), True
-
-        # 2. Try to restore from DB turns (survives Cloud Run restart)
-        if db is not None and db_session_id is not None:
-            try:
-                # Import here to avoid circular imports at module load time
-                from ..models.session import DialogueTurn  # noqa: PLC0415
-
-                turns = (
-                    db.query(DialogueTurn)
-                    .filter(DialogueTurn.learning_session_id == db_session_id)
-                    .order_by(DialogueTurn.turn_order)
-                    .all()
-                )
-
-                if turns:
-                    state = self._rebuild_state_from_turns(
-                        turns=turns,
-                        session_id=session_id,
-                        story_title=story_title,
-                        story_text=story_text,
-                        mispronounced_words=mispronounced_words,
-                        accuracy=accuracy,
-                        cpm=cpm,
-                        teacher_instructions=teacher_instructions,
-                        genre=genre,
-                        reading_strategy=reading_strategy,
-                    )
-                    _store.save(state, db=db, db_session_id=db_session_id)
-                    last_ai_turn = next(
-                        (t for t in reversed(state.conversation) if t["role"] == "ai"),
-                        None,
-                    )
-                    question = last_ai_turn["text"] if last_ai_turn else "請繼續回答上一個問題。"
-                    return AgentResponse(
-                        question=question,
-                        feedback=None,
-                        understood=None,
-                        understood_count=state.understood_count,
-                        required_count=self.REQUIRED_UNDERSTOOD,
-                        phase=state.current_phase,
-                        is_complete=state.understood_count >= self.REQUIRED_UNDERSTOOD,
-                    ), True
-            except (SQLAlchemyError, AttributeError, KeyError) as e:
-                logger.warning("Failed to restore session from DB turns (will start fresh): %s", e)
-
-        # 3. Fresh start — only Gemini call that's actually needed
-        state = SessionState(
-            session_id=session_id,
-            story_title=story_title,
-            story_text=story_text,
-            mispronounced_words=mispronounced_words,
-            accuracy=accuracy,
-            cpm=cpm,
-            teacher_instructions=teacher_instructions,
-            genre=genre,
-            reading_strategy=reading_strategy,
-        )
-
-        system_prompt = self._build_system_prompt(state)
-        contents = [
-            genai_types.Content(
-                role="user",
-                parts=[genai_types.Part(text="請開始提問。請提出第一個事實性問題。")],
-            )
-        ]
-
-        try:
-            result = await generate_structured_response(
-                system_prompt=system_prompt,
-                contents=contents,
-                response_schema=EVALUATION_SCHEMA,
-                task="socratic_question",
-            )
-            question = result.get("question", "這篇課文的主角是誰？")
-            phase = result.get("phase", "factual")
-        except Exception as e:
-            logger.warning("AI service error in start_session: %s", e)
-            question = "這篇課文的主角是誰？他（她）做了什麼事？"
-            phase = "factual"
-
-        state.current_phase = phase
-        state.conversation.append({"role": "ai", "text": question})
-        _store.save(state, db=db, db_session_id=db_session_id)
-
-        return AgentResponse(
-            question=question,
-            feedback=None,
-            understood=None,
-            understood_count=0,
-            required_count=self.REQUIRED_UNDERSTOOD,
-            phase=phase,
-            is_complete=False,
-        ), False
-
-    def _rebuild_state_from_turns(
-        self,
-        turns: list,
-        session_id: str,
-        story_title: str,
-        story_text: str,
-        mispronounced_words: list[str] | None,
-        accuracy: float | None,
-        cpm: float | None,
-        teacher_instructions: list[str] | None,
-        genre: str | None,
-        reading_strategy: str | None,
-    ) -> "SessionState":
-        """Rebuild in-memory SessionState from persisted DialogueTurn rows."""
-        state = SessionState(
-            session_id=session_id,
-            story_title=story_title,
-            story_text=story_text,
-            mispronounced_words=mispronounced_words,
-            accuracy=accuracy,
-            cpm=cpm,
-            teacher_instructions=teacher_instructions,
-            genre=genre,
-            reading_strategy=reading_strategy,
-        )
-
-        # Replay turns to reconstruct conversation and scores
-        for turn in turns:
-            state.conversation.append({"role": turn.role, "text": turn.text})
-            if turn.role == "student":
-                state.total_attempts += 1
-            elif turn.role == "feedback" and turn.is_correct:
-                state.understood_count += 1
-            if turn.phase:
-                state.current_phase = turn.phase
-
-        # Recalculate phase based on understood_count (mirrors process_answer logic)
-        if state.understood_count <= 1:
-            state.current_phase = "factual"
-        elif state.understood_count <= 3:
-            state.current_phase = "inferential"
-        else:
-            state.current_phase = "evaluative"
-
-        # Override with the phase of the latest turn if available
-        for turn in reversed(turns):
-            if turn.phase:
-                state.current_phase = turn.phase
-                break
-
-        return state
-
-    def clear_session(
-        self,
-        session_id: str,
-        db: Session | None = None,
-        db_session_id: int | None = None,
-    ) -> None:
-        """Remove a session from memory and optionally delete its DB turns.
-
-        Called when the student clicks '重新開始'.
-        """
-        # Remove from in-memory store
-        if session_id in _store._sessions:
-            del _store._sessions[session_id]
-
-        # Delete DB turns and dialogue_state snapshot so next mount starts fresh
-        if db is not None and db_session_id is not None:
-            try:
-                from ..models.session import DialogueTurn, LearningSession  # noqa: PLC0415
-
-                db.query(DialogueTurn).filter(
-                    DialogueTurn.learning_session_id == db_session_id
-                ).delete(synchronize_session=False)
-
-                ls = db.query(LearningSession).filter_by(id=db_session_id).first()
-                if ls is not None:
-                    ls.dialogue_state = None
-
-                db.commit()
-            except SQLAlchemyError as e:
-                logger.warning("Failed to delete DB dialogue turns on restart: %s", e)
-
-    async def process_answer(
-        self,
-        session_id: str,
-        student_answer: str,
-        db: Session | None = None,
-        db_session_id: int | None = None,
-    ) -> AgentResponse:
-        """Process a student's answer, evaluate understanding, return next question."""
-        # Rate limiting check
-        if _store.check_rate_limit(session_id):
-            raise ValueError("Rate limit exceeded. Please wait before sending another answer.")
-
-        # Input validation
-        if not student_answer or not student_answer.strip():
-            raise ValueError("Answer cannot be empty")
-        if len(student_answer) > self.MAX_ANSWER_LENGTH:
-            raise ValueError(f"Answer too long (max {self.MAX_ANSWER_LENGTH} characters)")
-        student_answer = student_answer.strip()
-
-        # Sanitize input to prevent prompt injection (Issue #270)
-        student_answer, _was_sanitized = sanitize_ai_input(
-            student_answer, user_id=session_id
-        )
-
-        state = _store.get(session_id, db=db, db_session_id=db_session_id)
-        if state is None:
-            raise ValueError(f"Session {session_id} not found or expired")
-
-        state.conversation.append({"role": "student", "text": student_answer})
-        state.total_attempts += 1
-
-        # Truncate conversation history to keep only last N turns
-        if len(state.conversation) > self.MAX_HISTORY_TURNS:
-            # Keep first turn (initial question) + last MAX_HISTORY_TURNS-1 turns
-            state.conversation = [state.conversation[0]] + state.conversation[-(self.MAX_HISTORY_TURNS - 1):]
-
-        system_prompt = self._build_system_prompt(state)
-
-        # Build Gemini contents from conversation
-        contents: list[genai_types.Content] = [
-            genai_types.Content(
-                role="user",
-                parts=[genai_types.Part(text="請開始提問。")],
-            )
-        ]
-        for turn in state.conversation:
-            role = "model" if turn["role"] == "ai" else "user"
-            contents.append(
-                genai_types.Content(
-                    role=role,
-                    parts=[genai_types.Part(text=turn["text"])],
-                )
-            )
-        # Ensure last message is from user
-        if contents[-1].role == "model":
-            contents.append(
-                genai_types.Content(
-                    role="user",
-                    parts=[genai_types.Part(text="請根據我的回答進行評估，然後繼續提問。")],
-                )
-            )
-
-        try:
-            result = await generate_structured_response(
-                system_prompt=system_prompt,
-                contents=contents,
-                response_schema=EVALUATION_SCHEMA,
-                task="socratic_agent_process",
-            )
-            understood = result.get("understood", False)
-            feedback = result.get("feedback", "")
-            question = result.get("question", "")
-            phase = result.get("phase", state.current_phase)
-            referenced_paragraph = result.get("referenced_paragraph")
-
-            # Validate referenced_paragraph bounds
-            num_paragraphs = len([p for p in state.story_text.split("\n") if p.strip()])
-            if referenced_paragraph is not None:
-                if not isinstance(referenced_paragraph, int) or referenced_paragraph < 1 or referenced_paragraph > num_paragraphs:
-                    logger.warning("Invalid referenced_paragraph %s (valid: 1-%d), resetting to None", referenced_paragraph, num_paragraphs)
-                    referenced_paragraph = None
-
-            # Validate phase
-            if phase not in PHASE_ORDER:
-                logger.warning("Invalid phase '%s', using current: %s", phase, state.current_phase)
-                phase = state.current_phase
-
-            # Validate question is non-empty
-            if not question or not question.strip():
-                question = _fallback_question(state)
-
-            state.consecutive_errors = 0  # Reset on success
-
-        except Exception as e:
-            state.consecutive_errors += 1
-            logger.warning(
-                "AI service error in process_answer (attempt %d/%d): %s",
-                state.consecutive_errors,
-                self.MAX_CONSECUTIVE_ERRORS,
-                e,
-                extra={
-                    "event": "socratic_ai_error",
-                    "consecutive_errors": state.consecutive_errors,
-                    "max_consecutive_errors": self.MAX_CONSECUTIVE_ERRORS,
-                    "session_id": session_id,
-                    "error": str(e),
-                },
-            )
-
-            if state.consecutive_errors >= self.MAX_CONSECUTIVE_ERRORS:
-                logger.error(
-                    "Circuit breaker triggered: %d consecutive AI errors for session %s",
-                    state.consecutive_errors,
-                    session_id,
-                    extra={
-                        "event": "circuit_breaker_triggered",
-                        "consecutive_errors": state.consecutive_errors,
-                        "session_id": session_id,
-                    },
-                )
-                _store.save(state, db=db, db_session_id=db_session_id)
-                raise RuntimeError("AI 服務暫時無法使用，請稍後再試。") from e
-
-            understood = False  # Don't auto-pass on error; re-ask instead
-            feedback = "讓我再想一下，請你再回答一次好嗎？"
-            question = _fallback_question(state)
-            phase = state.current_phase
-            referenced_paragraph = None
-
-        if understood:
-            state.understood_count += 1
-            referenced_paragraph = None  # Clear on correct answer
-            # Advance phase based on understood count.
-            # Phase sets the prompt for the NEXT question, so offset by 1:
-            # count 0-1 → factual (Q1-Q2), 2-3 → inferential (Q3-Q4), 4+ → evaluative (Q5)
-            if state.understood_count <= 1:
-                state.current_phase = "factual"
-            elif state.understood_count <= 3:
-                state.current_phase = "inferential"
-            else:
-                state.current_phase = "evaluative"
-            phase = state.current_phase
-
-        is_complete = state.understood_count >= self.REQUIRED_UNDERSTOOD
-
-        if not is_complete:
-            state.conversation.append({"role": "ai", "text": question})
-
-        _store.save(state, db=db, db_session_id=db_session_id)
-
-        return AgentResponse(
-            question=question,
-            feedback=feedback,
-            understood=understood,
-            understood_count=state.understood_count,
-            required_count=self.REQUIRED_UNDERSTOOD,
-            phase=phase,
-            is_complete=is_complete,
-            referenced_paragraph=referenced_paragraph,
-        )
-
-
-def _fallback_question(state: SessionState) -> str:
-    """Return a pre-written question when AI is unavailable."""
-    fallback = {
-        "factual": "這篇課文的主角是誰？他（她）做了什麼事？",
-        "inferential": "為什麼主角要這樣做？你覺得他的理由合理嗎？",
-        "evaluative": "讀完這篇課文，你有什麼感想？如果是你，你會怎麼做？",
-    }
-    return fallback.get(state.current_phase, fallback["factual"])
-
-
-# Module-level agent singleton
-socratic_agent = SocraticAgent()
+"""Compatibility shim — socratic agent now lives in services/socratic/."""
+from .socratic import SocraticAgent, AgentResponse, SessionState, SessionStore, socratic_agent  # noqa: F401
